For some reason, subversion decided not to update header in previous commit.
git-svn-id: http://svn.mangler.org/mangler/trunk@54 f654edcb-a5fb-4f28-a0e7-68c327795dae

+++ libventrilo3/libventrilo3_message.h
@@ -59,7 +59,7 @@ typedef struct _v3_net_message_0x00 {/*{{{*/
      *  Message Type 0x00 is only used during the login phase as a part of the
      *  key exchange.  The client sends 64 bytes of random data to the server.
      */
-    uint16_t type;
+    uint32_t type;
     char version[16];
     char salt1[32];
     char salt2[32];
